Removed CommandManager | Overhauls | Added signals | Added to Fc default config | Fcs delegating more to block managers
Overhauls:
- Command and Block ids are now bytes
  - Values set relative to their containers
    - Blocks now set their Commands' unique ids
- Block IsExecuting func turned into property
- Unique Block names handled by UniqueKeyGenerator
- Split from ICommandSource: IHasCommands and ICommandRemovable
- BlockLogicManager no longer an ICommandSource
  - It does expect you to give it an IBlockManager so it can properly track and execute things, though
- Switched out some linq statements for the sake of Unity 2022.3
- Removed unnecessary usings

+++ Assets/Hyphlow/Editor/Scripts/FlowchartWindow/FcWindowComponents/BlockModuleDispatcher.cs
@@ -151,10 +151,10 @@ namespace AtMycelia.Hyphlow.EditorUtils.FcWindow
         public void NotifyPreMultiBlockCut(IList<IBlock> blocks) =>
             Broadcast<IPreMultiBlockCutResponder>(res => res.OnPreMultiBlockCut(blocks));
 
-        public void NotifyPostBlockCut(ushort blockId) =>
+        public void NotifyPostBlockCut(byte blockId) =>
             Broadcast<IPostBlockCutResponder>(res => res.OnPostBlockCut(blockId));
 
-        public void NotifyPostMultiBlockCut(IList<ushort> blockIds) =>
+        public void NotifyPostMultiBlockCut(IList<byte> blockIds) =>
             Broadcast<IPostMultiBlockCutResponder>(res => res.OnPostMultiBlockCut(blockIds));
 
         public void NotifyBlockCreated(IBlock block) =>
@@ -166,10 +166,10 @@ namespace AtMycelia.Hyphlow.EditorUtils.FcWindow
         public void NotifyPreMultiBlockDeleted(IList<IBlock> blocks) =>
             Broadcast<IPreBlockDeletionResponder>(res => res.OnPreBlockDeletion(blocks));
 
-        public void NotifyPostBlockDeleted(ushort blockId) =>
+        public void NotifyPostBlockDeleted(byte blockId) =>
             Broadcast<IPostBlockDeletionResponder>(res => res.OnPostBlockDeletion(blockId));
 
-        public void NotifyPostMultiBlockDeleted(IList<ushort> blockIds) =>
+        public void NotifyPostMultiBlockDeleted(IList<byte> blockIds) =>
             Broadcast<IPostMultiBlockDeletionResponder>(res => res.OnPostMultiBlockDeletion(blockIds));
 
         public void NotifyBlockClicked(IBlock block, Event evt) =>
